Add files via upload
The ABC company CEO is looking to add new items to the menu in the fast-food department. The CEO needed help in determining the promotion campaign for the new products. Objective: quickly analyze test market campaigns based on responses, revenue, and other key metrics. Predict who will respond to which campaign by which channel and why. Increase the likelihood of responses and quality of leads in future campaigns. A different promotion is used at each location, and the weekly sales of the new item are recorded for the first four weeks.